# src/Mimo.Admin.Api.Sdk/openapi/mimo-admin-api.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Mimo.Admin.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5098/"
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Autentica a un super administrador y emite un token JWT.",
        "operationId": "SuperAdminLogin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/auth/setup": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Crea el primer super administrador de la plataforma. Solo funciona si aún no existe ninguno.",
        "operationId": "SetupSuperAdmin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetupSuperAdminRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/tenants": {
      "get": {
        "tags": [
          "Tenants"
        ],
        "summary": "Lista todos los tenants con paginación opcional.",
        "operationId": "ListTenants",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "isActive",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfTenantResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tenants"
        ],
        "summary": "Registra un nuevo tenant y aprovisiona su esquema en la BD.",
        "operationId": "CreateTenant",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTenantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "409": {
            "description": "Conflict"
          }
        }
      },
      "put": {
        "tags": [
          "Tenants"
        ],
        "summary": "Actualiza nombre, plan y estado activo de un tenant (query: id).",
        "operationId": "UpdateTenant",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Tenants"
        ],
        "summary": "Desactiva un tenant / borrado lógico (query: id).",
        "operationId": "DeactivateTenant",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/tenants/detail": {
      "get": {
        "tags": [
          "Tenants"
        ],
        "summary": "Obtiene un tenant por su ID (query: id).",
        "operationId": "GetTenant",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/plans": {
      "get": {
        "tags": [
          "Plans"
        ],
        "summary": "Lista los planes del catálogo.",
        "operationId": "ListPlans",
        "parameters": [
          {
            "name": "isActive",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Plans"
        ],
        "summary": "Crea un plan en el catálogo.",
        "operationId": "CreatePlan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Plans"
        ],
        "summary": "Actualiza nombre, descripción y estado de un plan (query: code).",
        "operationId": "UpdatePlan",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Plans"
        ],
        "summary": "Desactiva un plan / borrado lógico (query: code).",
        "operationId": "DeactivatePlan",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/plans/detail": {
      "get": {
        "tags": [
          "Plans"
        ],
        "summary": "Obtiene un plan por código (query: code).",
        "operationId": "GetPlan",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/ai/connectors": {
      "get": {
        "tags": [
          "AiConnectors"
        ],
        "summary": "Lista los conectores de IA.",
        "operationId": "ListAiConnectors",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "AiConnectors"
        ],
        "summary": "Registra un nuevo conector de IA.",
        "operationId": "CreateAiConnector",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAiConnectorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "AiConnectors"
        ],
        "summary": "Actualiza nombre y configuración de un conector (query: id).",
        "operationId": "UpdateAiConnector",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAiConnectorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/ai/connectors/activate": {
      "post": {
        "tags": [
          "AiConnectors"
        ],
        "summary": "Marca un conector como el activo de la plataforma (query: id).",
        "operationId": "ActivateAiConnector",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/ai/plan-policies": {
      "get": {
        "tags": [
          "AiPlanPolicies"
        ],
        "summary": "Lista las políticas de IA por plan.",
        "operationId": "ListAiPlanPolicies",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "AiPlanPolicies"
        ],
        "summary": "Crea o actualiza la política de IA de un plan (query: planCode).",
        "operationId": "UpsertAiPlanPolicy",
        "parameters": [
          {
            "name": "planCode",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertAiPlanPolicyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "AiPlanPolicies"
        ],
        "summary": "Elimina la política de IA de un plan (query: planCode).",
        "operationId": "DeleteAiPlanPolicy",
        "parameters": [
          {
            "name": "planCode",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/ai/plan-policies/detail": {
      "get": {
        "tags": [
          "AiPlanPolicies"
        ],
        "summary": "Obtiene la política de un plan (query: planCode).",
        "operationId": "GetAiPlanPolicy",
        "parameters": [
          {
            "name": "planCode",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/ai/usage/summary": {
      "get": {
        "tags": [
          "AiUsage"
        ],
        "summary": "Resume el uso de IA por tenant en un periodo (query: from, to, tenantId; por defecto el mes actual).",
        "operationId": "GetAiUsageSummary",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "tenantId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AiConnectorSettingsDto": {
        "required": [
          "apiKey",
          "baseUrl",
          "chatModel",
          "embeddingModel"
        ],
        "type": "object",
        "properties": {
          "apiKey": {
            "type": "string"
          },
          "baseUrl": {
            "type": "string"
          },
          "chatModel": {
            "type": "string"
          },
          "embeddingModel": {
            "type": "string"
          },
          "temperature": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "format": "float",
            "default": 0.7
          },
          "maxTokens": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "format": "int32",
            "default": 1024
          }
        }
      },
      "CreateAiConnectorRequest": {
        "required": [
          "provider",
          "displayName",
          "settings"
        ],
        "type": "object",
        "properties": {
          "provider": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "settings": {
            "$ref": "#/components/schemas/AiConnectorSettingsDto"
          }
        }
      },
      "CreatePlanRequest": {
        "required": [
          "code",
          "name"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CreateTenantRequest": {
        "required": [
          "name",
          "slug",
          "plan",
          "adminEmail",
          "adminPassword",
          "adminFullName",
          "adminAlias"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "plan": {
            "type": "string"
          },
          "adminEmail": {
            "type": "string"
          },
          "adminPassword": {
            "type": "string"
          },
          "adminFullName": {
            "type": "string"
          },
          "adminAlias": {
            "type": "string"
          }
        }
      },
      "LoginRequest": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "LoginResponse": {
        "required": [
          "accessToken",
          "expiresAtUtc",
          "id",
          "email",
          "fullName",
          "roles"
        ],
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "expiresAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "PagedResultOfTenantResponse": {
        "required": [
          "items",
          "totalCount",
          "page",
          "pageSize"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TenantResponse"
            }
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "format": "int32"
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "format": "int32"
          },
          "totalPages": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "format": "int32"
          },
          "hasPrevious": {
            "type": "boolean"
          },
          "hasNext": {
            "type": "boolean"
          }
        }
      },
      "SetupSuperAdminRequest": {
        "required": [
          "email",
          "password",
          "fullName"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          }
        }
      },
      "TenantResponse": {
        "required": [
          "id",
          "name",
          "slug",
          "plan",
          "isActive",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "plan": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UpdateAiConnectorRequest": {
        "required": [
          "displayName",
          "settings"
        ],
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string"
          },
          "settings": {
            "$ref": "#/components/schemas/AiConnectorSettingsDto"
          }
        }
      },
      "UpdatePlanRequest": {
        "required": [
          "name",
          "description"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean",
            "default": true
          }
        }
      },
      "UpdateTenantRequest": {
        "required": [
          "name",
          "plan",
          "isActive"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "plan": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
      "UpsertAiPlanPolicyRequest": {
        "required": [
          "allowedProviders"
        ],
        "type": "object",
        "properties": {
          "allowedProviders": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "monthlyRequestQuota": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "format": "int32",
            "default": 0
          },
          "monthlyTokenQuota": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "format": "int64",
            "default": 0
          },
          "isActive": {
            "type": "boolean",
            "default": true
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Auth"
    },
    {
      "name": "Tenants"
    },
    {
      "name": "Plans"
    },
    {
      "name": "AiConnectors"
    },
    {
      "name": "AiPlanPolicies"
    },
    {
      "name": "AiUsage"
    }
  ]
}